Split scrape command
