+++ cmd/util-db/main.go
@@ -17,20 +17,19 @@
 package main
 
 import (
+	"log"
+	"os"
+
 	"github.com/0xsoniclabs/aida/cmd/util-db/clone"
 	"github.com/0xsoniclabs/aida/cmd/util-db/compact"
 	"github.com/0xsoniclabs/aida/cmd/util-db/generate"
 	"github.com/0xsoniclabs/aida/cmd/util-db/info"
 	"github.com/0xsoniclabs/aida/cmd/util-db/merge"
 	"github.com/0xsoniclabs/aida/cmd/util-db/metadata"
+	"github.com/0xsoniclabs/aida/cmd/util-db/primer"
+	"github.com/0xsoniclabs/aida/cmd/util-db/scrape"
 	"github.com/0xsoniclabs/aida/cmd/util-db/update"
 	"github.com/0xsoniclabs/aida/cmd/util-db/validate"
-	"log"
-	"os"
-
-	"github.com/0xsoniclabs/aida/cmd/util-db/db"
-	"github.com/0xsoniclabs/aida/cmd/util-db/primer"
-
 	"github.com/urfave/cli/v2"
 )
 
@@ -50,7 +49,7 @@ var UtilDbApp = cli.App{
 		&generate.EthereumGenesisCommand,
 		&generate.DeletedAccountsCommand,
 		&generate.DbHashCommand,
-		&db.ScrapeCommand,
+		&scrape.Command,
 		&validate.Command,
 
 		//Priming only
