+++ cmd/util-db/scrape/scrape.go
@@ -1,20 +1,4 @@
-// Copyright 2024 Fantom Foundation
-// This file is part of Aida Testing Infrastructure for Sonic
-//
-// Aida is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// Aida is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with Aida. If not, see <http://www.gnu.org/licenses/>.
-
-package db
+package scrape
 
 import (
 	"fmt"
@@ -25,8 +9,8 @@ import (
 	"github.com/urfave/cli/v2"
 )
 
-var ScrapeCommand = cli.Command{
-	Action:    scrapePrepare,
+var Command = cli.Command{
+	Action:    scrapeAction,
 	Name:      "scrape",
 	Usage:     "Stores state hashes into TargetDb for given range",
 	ArgsUsage: "<blockNumFirst> <blockNumLast>",
@@ -38,8 +22,8 @@ var ScrapeCommand = cli.Command{
 	},
 }
 
-// scrapePrepare stores state hashes into Target for given range
-func scrapePrepare(ctx *cli.Context) error {
+// scrapeAction stores state hashes into Target for given range
+func scrapeAction(ctx *cli.Context) error {
 	cfg, argErr := utils.NewConfig(ctx, utils.BlockRangeArgs)
 	if argErr != nil {
 		return argErr
